separate OS selection boxes to use as a stand alone component for validators as well
